requirements.txt
aniso8601==9.0.1
aws==0.2.5
awscli==1.32.61
bcrypt==4.1.2
blinker==1.7.0
boto==2.49.0
boto3==1.34.61
botocore==1.34.61
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
confluent-kafka==2.3.0
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_daq==0.5.0
decorator==5.1.1
deepdiff==6.7.1
defusedxml==0.7.1
Deprecated==1.2.14
dictdiffer==0.9.0
distlib==0.3.8
docker==7.0.0
docutils==0.16
exceptiongroup==1.2.0
fabric==3.2.2
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
fonttools==4.51.0
furl==2.1.3
genson==1.2.2
gitdb==4.0.11
GitPython==3.1.42
identify==2.5.35
idna==3.6
importlib_metadata==7.1.0
iniconfig==2.0.0
invoke==2.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
jira==3.8.0
jmespath==1.0.1
json2html==1.3.0
kaleido==0.2.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
mpld3==0.5.10
nest-asyncio==1.6.0
nodeenv==1.8.0
numpy==1.26.4
oauthlib==3.2.2
ordered-set==4.1.0
orderedmultidict==1.0.1
packaging==24.0
pandas==2.2.1
paramiko==3.4.0
pdfkit==1.0.0
pillow==10.3.0
pipenv==2023.12.1
platformdirs==4.2.0
plotly==5.19.0
pluggy==1.4.0
pre-commit==3.6.2
prettytable==3.10.0
pyasn1==0.5.1
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.1.2
pytest==8.0.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retrying==1.3.4
rsa==4.7.2
rstr==3.2.2
s3transfer==0.10.0
six==1.16.0
smmap==5.0.1
tenacity==8.2.3
tomli==2.0.1
typing_extensions==4.11.0
tzdata==2024.1
virtualenv==20.25.1
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.18.1
PyPDF2~=3.0.1
reportlab~=4.2.0
mysql-connector-python==8.4.0
pyscheduler==0.1.0
SQLAlchemy==2.0.30
cattrs
jira==3.8.0
jira==3.8.0
apscheduler
jsonschema
genson
